Upgrade to level of openafs-1.6.3pre3

+++ src/ubik/remote.c
@@ -636,12 +636,15 @@ SDISK_SendFile(struct rx_call *rxcall, afs_int32 file, afs_int32 index,
 
     /* send the file back to the requester */
 
-    if ((code = ubik_CheckAuth(rxcall))) {
+    pbuffer[0] = '\0';
+    dbase = ubik_dbase[index];
+    if (!dbase) {
+	code = ENOENT;
 	goto failed;
     }
 
-    if (!ubik_dbase[index]) {
-	code = ENOENT;
+    if ((code = ubik_CheckAuth(rxcall))) {
+	DBHOLD(dbase);
 	goto failed;
     }
 
@@ -665,7 +668,6 @@ SDISK_SendFile(struct rx_call *rxcall, afs_int32 file, afs_int32 index,
 	goto failed;
     }
 
-    dbase = ubik_dbase[index];
     DBHOLD(dbase);
 
     /* abort any active trans that may scribble over the database */
@@ -706,7 +708,6 @@ SDISK_SendFile(struct rx_call *rxcall, afs_int32 file, afs_int32 index,
 #endif
 	code = rx_Read(rxcall, tbuffer, tlen);
 	if (code != tlen) {
-	    DBRELE(dbase);
 	    ubik_dprint("Rx-read length error=%d\n", code);
 	    code = BULK_ERROR;
 	    close(fd);
@@ -719,7 +720,6 @@ SDISK_SendFile(struct rx_call *rxcall, afs_int32 file, afs_int32 index,
 	pass++;
 #endif
 	if (code != tlen) {
-	    DBRELE(dbase);
 	    ubik_dprint("write failed error=%d\n", code);
 	    code = UIOERROR;
 	    close(fd);
@@ -767,11 +767,12 @@ SDISK_SendFile(struct rx_call *rxcall, afs_int32 file, afs_int32 index,
 #else
     LWP_NoYieldSignal(&dbase->version);
 #endif
-    DBRELE(dbase);
-  failed:
+
+failed:
     if (code) {
 #ifndef OLD_URECOVERY
-	unlink(pbuffer);
+	if (pbuffer[0] != '\0')
+	    unlink(pbuffer);
 	/* Failed to sync. Allow reads again for now. */
 	if (dbase != NULL) {
 	    tversion.epoch = epoch;
@@ -784,6 +785,7 @@ SDISK_SendFile(struct rx_call *rxcall, afs_int32 file, afs_int32 index,
     } else {
 	ubik_print("Ubik: Synchronize database %s completed\n", dbase->pathName);
     }
+    DBRELE(dbase);
     return code;
 }
 
@@ -856,10 +858,9 @@ SDISK_UpdateInterfaceAddr(struct rx_call *rxcall,
     /* if (probableMatch) */
     /* inconsistent addresses in CellServDB */
     if (!probableMatch || found) {
-	ubik_print("Inconsistent Cell Info from server: ");
+	ubik_print("Inconsistent Cell Info from server:\n");
 	for (i = 0; i < UBIK_MAX_INTERFACE_ADDR && inAddr->hostAddr[i]; i++)
-	    ubik_print("%s ", afs_inet_ntoa_r(htonl(inAddr->hostAddr[i]), hoststr));
-	ubik_print("\n");
+	    ubik_print("... %s\n", afs_inet_ntoa_r(htonl(inAddr->hostAddr[i]), hoststr));
 	fflush(stdout);
 	fflush(stderr);
 	printServerInfo();
@@ -870,15 +871,9 @@ SDISK_UpdateInterfaceAddr(struct rx_call *rxcall,
     for (i = 1; i < UBIK_MAX_INTERFACE_ADDR; i++)
 	ts->addr[i] = htonl(inAddr->hostAddr[i]);
 
-    msg = malloc(4096);
-    sprintf(msg, "ubik: A Remote Server has addresses: ");
-    for (i = 0; i < UBIK_MAX_INTERFACE_ADDR && ts->addr[i]; i++) {
-	sprintf(ipstr, "%s, ", afs_inet_ntoa_r(ts->addr[i], hoststr));
-	strcat(msg, ipstr);
-    }
-    msg[strlen(msg)-2] = 0;
-    ubik_print("%s\n", msg);
-    free(msg);
+    ubik_print("ubik: A Remote Server has addresses:\n");
+    for (i = 0; i < UBIK_MAX_INTERFACE_ADDR && ts->addr[i]; i++)
+	ubik_print("... %s\n", afs_inet_ntoa_r(ts->addr[i], hoststr));
 
     return 0;
 }
@@ -890,13 +885,12 @@ printServerInfo(void)
     int i, j = 1;
     char hoststr[16];
 
-    ubik_print("Local CellServDB:");
+    ubik_print("Local CellServDB:\n");
     for (ts = ubik_servers; ts; ts = ts->next, j++) {
-	ubik_print("Server %d: ", j);
+	ubik_print("  Server %d:\n", j);
 	for (i = 0; (i < UBIK_MAX_INTERFACE_ADDR) && ts->addr[i]; i++)
-	    ubik_print("%s ", afs_inet_ntoa_r(ts->addr[i], hoststr));
+	    ubik_print("  ... %s\n", afs_inet_ntoa_r(ts->addr[i], hoststr));
     }
-    ubik_print("\n");
 }
 
 afs_int32
